[refactor] Splits off part of apiController into smaller files  - Adds orgsController and usersController  - Reorganizes server directory
