+++ .gitignore
@@ -29,6 +29,6 @@ build/Release
 node_modules
 
 server/config/config.js
-server/utils/dbUtils/chimeTeam.sql
+server/db/chimeTeam.sql
 
 .DS_Store
