+++ circle.yml
@@ -2,3 +2,4 @@ machine:
   environment:
     DATABASE_TEST_URL: postgres://ubuntu@localhost/circle_test
     SESSION_SECRET: 'foobarbydonthurtme'
+    TEST: true
