+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "bcrypt-nodejs": "0.0.3",
+    "bluebird": "^2.9.25",
     "body-parser": "~1.12.0",
     "bootstrap": "^3.3.4",
     "browserify": "~10.1.0",
