--- server/data/testData.js
@@ -1,45 +1,0 @@
-/* This is data for testing when the server is not connected
- * to the database.
- */
- var orgInfo = {
-  name: 'Makersquare',
-  members: [
-    {
-      id: 1,
-      first_name: 'Bobby',
-      last_name: 'Hill',
-      photo: 'http://i.imgur.com/KQjdi0i.jpg',
-      title: 'Student'
-    },
-    {
-      id: 2,
-      first_name: 'Hank',
-      last_name: 'Hill',
-      photo: 'http://i.imgur.com/sY47Zl6.jpg',
-      title: 'Propane Salesman'
-    },
-    {
-      id: 3,
-      first_name: 'Luanne',
-      last_name: 'Platter',
-      photo: 'http://i.imgur.com/hUoJph9.jpg',
-      title: 'Beautician'
-    },
-    {
-      id: 4,
-      first_name: 'Dale',
-      last_name: 'Gribble',
-      photo: 'http://i.imgur.com/FJw1Nbb.jpg',
-      title: 'Exterminator'
-    },
-    {
-      id: 5,
-      first_name: 'Peggy',
-      last_name: 'Hill',
-      photo: 'http://i.imgur.com/gnzS7G5.jpg',
-      title: 'Homemaker'
-    }
-  ]
-};
-
-exports.data = orgInfo;
