+++ server/ping/smsUtils.js
@@ -4,8 +4,8 @@ var _ = require('underscore');
 // Create reusable client object
 if (!process.env.TEST) {
   var client = new voicejs.Client({
-    email: process.env.GMAIL_EMAIL || require('../../config/config').gmail.email,
-    password: process.env.GMAIL_PASSWORD || require('../../config/config').gmail.password
+    email: process.env.GMAIL_EMAIL || require('../config/config').gmail.email,
+    password: process.env.GMAIL_PASSWORD || require('../config/config').gmail.password
   });
 }
